Refactors ShuffleDeck class to be generic.
Replaces ArrayList implementation with generics to avoid having to
cast return values back to the right types when using a deck.

Adds deck and RNG state tracking for save/restore. Untested. Ideally,
will allow for the game to remember a deck's settings and resume
where it left off with the same card order and RNG seed.