Converts mobile control from gyro to compass.
Gyro had too much drift for long play sessions.
This converts the controller over to the compass.
The compass is a little jittery, but works fine
for an audio-only game where you can't see the
jitter.

Also adds an improved GyroCam script based on
a forum post. Still drifty, though.

Adds a 4-finger reset on mobile. Touch 4 fingers
for 4 seconds and the game will reset to the
opening instructional monologue.